Update crud.php
This is the main stream. Here is the event-loop.
As an CRUD operation is needed, it is directly called and receive data or an boolean status.

Why not using routes?
Because this is to be used in situations where the access to database data is needed in to satisfy an especific routine that needs this data, or must save it.